Added instructions for setting up the AP on the RPi3
Instructions for setting up a system where an AP that can be accessed by the RPiZeros is set up on the RPi3s without losing connection to the internet. This effectively means that the RPiZeros can access the internet through the RPi3s which act as a hub/router/repeater.