Automatic updating changes
Moved updates to their own script to remove the need for a full 
reinstall of everything. Packages reinstalled or updated are only those 
which have been added since the pilot, ie the sharepoint packages, the 
adafruit packages, and the RTC installation.
Re-enabled the tests in opc_install as that script would now only be 
used when doing a first-time setup.
Changed rc.local git pulls so that if there are changes in the repo 
detected on startup, a reboot will be triggered. This means that changes 
to rc.local can be brought in straight away rather than on the next 
reboot, whenever that would be.